refactor(#800): split commands/stats.rs by moving tests to sibling
Move 4,216-line commands/stats.rs into commands/stats/ directory:

- stats/mod.rs:   2,499 LOC (production code)
- stats/tests.rs: 1,714 LOC (unit tests)

Production module sits just under the 2,500 LOC limit. `super::*`
preserves test access to private items. No behavior changes.

`cargo test -p budi-cli` green.
